+++ src/main/java/com/graviteesource/policy/threatprotection/json/JsonException.java
@@ -1,0 +1,24 @@
+package com.graviteesource.policy.threatprotection.json;
+
+/**
+ * @author Jeoffrey HAEYAERT (jeoffrey.haeyaert at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+public class JsonException extends Exception {
+
+    private String key;
+
+    public JsonException(String key, String message) {
+        super(message);
+        this.key = key;
+    }
+
+    public JsonException(String key, String message, Exception e) {
+        super(message, e);
+        this.key = key;
+    }
+
+    public String getKey() {
+        return key;
+    }
+}
