+++ src/main/resources/plugin.properties
@@ -2,7 +2,7 @@ id=json-threat-protection
 name=JSON Threat Protection
 version=${project.version}
 description=${project.description}
-class=io.gravitee.policy.threatprotection.json.JsonThreatProtectionPolicy
+class=com.graviteesource.policy.threatprotection.json.JsonThreatProtectionPolicy
 type=policy
 category=security
 icon=json-threat-protection.svg
