+++ src/test/java/com/graviteesource/policy/threatprotection/json/JsonThreatProtectionPolicyTest.java
@@ -1,21 +1,6 @@
-/**
- * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.gravitee.policy.threatprotection.json;
+package com.graviteesource.policy.threatprotection.json;
 
-import static io.gravitee.policy.threatprotection.json.JsonThreatProtectionPolicy.*;
+import static com.graviteesource.policy.threatprotection.json.JsonThreatProtectionPolicy.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.*;
